+++ assets/mixins/index.js
@@ -5,5 +5,6 @@ module.exports = {
   formInput: require('./form-input.mixin.jsx'),
   form: require('./form.mixin'),
   pagination: require('./pagination.mixin'),
-  tooltip: require('./tooltip.mixin.jsx')
+  tooltip: require('./tooltip.mixin.jsx'),
+  validator: require('./validator.mixin')
 };